Refactor Test Suite
